Prototype: decouple Smack from XmppProtocolProvider
This commit adds an Adapter, that allows the XmppProtocolProvider to be implementation agnostic (and no longer depend on Smack). Looking at the existing Javadoc, this is how this class was designed to operate in the first place.

The Adapter implementation is highly experimental and untested at the time of this commit. This should be regarded as a prototype.

By decoupling Smack from the provider, other implementations become feasible. The end objective of this thread is to be able to serve Jicofo's functionality as server-sided functionality, rather than a client-sided functionality.